+++ get-config.js
@@ -24,9 +24,10 @@ function getGlobalConfig({
 	} = {}
 ) {
 	// Locate ESLint module resolver file.
-	const eslintResolverPath = search
-		.map(dir => path.join(dir, 'eslint', 'lib', 'shared', 'relative-module-resolver.js'))
-		.find(dir => fs.existsSync(dir));
+	const eslintResolverPath = [
+		...search.map(dir => path.join(dir, 'eslint', 'node_modules', '@eslint', 'eslintrc', 'lib', 'shared', 'relative-module-resolver.js')),
+		...search.map(dir => path.join(dir, 'eslint', 'lib', 'shared', 'relative-module-resolver.js'))
+	].find(dir => fs.existsSync(dir));
 	if (eslintResolverPath) {
 		supportGlobalResolving(eslintResolverPath, search);
 	}