+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact proxy ESLint config:
 
+## [unreleased]
+
+### Fixed
+
+- ESLint module resolver path to resolve correctly with the latest eslint.
+
 ## [1.0.1]
 
 ### Fixed